update order detail order list cart item 70%

+++ database/migrations/2023_09_23_091854_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('picture')->nullable();
             $table->string('name');
             $table->string('slug')->unique();
-            $table->double('price');
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
